Added what I have for course fitness and ES mods.
Adding what I have for the course schedule resolver (uses fitness) and the expert system. The course schedule resolver is mostly complete but does not have any concrete rules associated with it. I am reworking the expert system because I realized a flaw in my general approach--the new ES will work on a complete path to graduation in a single call.